refactor(chalice): use CH everywhere to get error's details
