+++ api/chalicelib/core/errors/__init__.py
@@ -1,8 +1,6 @@
 import logging
 
-from decouple import config
-
 logger = logging.getLogger(__name__)
 
 from . import errors_ch as errors
-
+from . import errors_details_ch as errors_details
