+++ api/chalicelib/core/sessions/sessions_replay.py
@@ -3,7 +3,7 @@ from chalicelib.core import metadata, assist, canvas, user_testing
 from chalicelib.core.issues import issues
 from chalicelib.core.events import events, events_mobile
 from . import sessions_mobs, sessions_devtool
-from chalicelib.core.errors.modules import errors_helper
+from chalicelib.core.errors import helper as errors_helper
 from chalicelib.utils import pg_client, helper
 from chalicelib.core.modules import MOB_KEY, get_file_key
